% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Christensen21.R
\name{Christensen}
\alias{Christensen}
\title{Sample size to acquire a given width by way of the confidence interval for the limits of agreement proposed in Christensen et al 2021 paper}
\usage{
Christensen(sigmab, sigmaw, ew, gamma = 0.95, alpha = 0.05)
}
\arguments{
\item{sigmab}{The standard deviation between individuals.}

\item{sigmaw}{The standard deviation within individuals.}

\item{ew}{The clinical acceptable width of the confidence interval.}

\item{gamma}{Coverage of the LOA. Default is set to 0.95.}

\item{alpha}{alpha level of the confidence interval. Default is set to 0.05.}
}
\value{
Factor variable
}
\description{
Computes the necessary sample size to achieve a given precision for the LOA by way of the confidence interval presented in 'The appropriateness of Bland-Altman's approximate confidence intervals for limits of agreement' by Christensen et al.
}
\examples{
See *my paper*
}
