% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jan18.R
\name{Jan18}
\alias{Jan18}
\title{Sample size to acquire a given width by way of the confidence interval for the limits of agreement proposed in Jan and Shiehs 2018 paper}
\usage{
Jan18(sigma, ew, alpha = 0.05, mu = 0, gamma = 0.95)
}
\arguments{
\item{sigma}{The standard deviation of the differences.}

\item{ew}{The clinical acceptable precision of the confidence interval.}

\item{alpha}{alpha level of the confidence interval. Default is set to 0.05.}

\item{mu}{Mean of the differences. Default is set to 0.}

\item{gamma}{Coverage of the LOA. Default is set to 0.95.}
}
\value{
Factor variable
}
\description{
Computes the necessary sample size to achieve a given precision for the LOA by way of the confidence interval presented in 'The appropriateness of Bland-Altman's approximate confidence intervals for limits of agreement' by Jan and Shieh.
}
\examples{
See *my paper*
}
