% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jan18assurance.R
\name{Jan18assurance}
\alias{Jan18assurance}
\title{Sample size for the agreement test proposed in Jan and Shieh's 2018 paper}
\usage{
Jan18assurance(sigma, delta, ap, me = 0, alpha = 0.05, gamma = 0.95)
}
\arguments{
\item{sigma}{The standard deviation of the differences.}

\item{delta}{The clinical acceptable width of the LOA.}

\item{ap}{The assurance probability.}

\item{me}{The mean of the differences. Default is set to 0.}

\item{alpha}{Significance level. Default is set to 0.05.}

\item{gamma}{Coverage of the LOA. Default is set to 0.95.}
}
\value{
Factor variable
}
\description{
Computes the necessary sample size to achieve a given width by way of the formula presented in 'Assessing Agreement Between Two Methods of Quantitative Measurements: Exact Test Procedure and Sample Size Calculation' by Jan and Shieh.
}
\examples{
See *my paper*
}
