% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lin98.R
\name{Lin98}
\alias{Lin98}
\title{Sample size for the agreement test proposed in Lin's 1998 paper}
\usage{
Lin98(delta, sigma, alpha = 0.05, beta = 0.2, gamma = 0.95)
}
\arguments{
\item{delta}{The clinical acceptable width of the LOA.}

\item{sigma}{The standard deviation of the differences.}

\item{alpha}{Significance level. Default is set to 0.05.}

\item{beta}{Type II error. Default is set to 0.2.}

\item{gamma}{Coverage of the LOA. Default is set to 0.95.}
}
\value{
Factor variable
}
\description{
Computes the necessary sample size to achieve a given power by way of the formula presented in 'Evaluation of statistical equivalence using limits of agreement and associated sample size calculation' by Lin et al.
}
\examples{
See *my paper*

}
