% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lu16.R
\name{Lu16}
\alias{Lu16}
\title{Sample size for the agreement test proposed in Lu's 2016 paper}
\usage{
Lu16(delta, sigma, me = 0, alpha = 0.05, beta = 0.2, gamma = 0.95)
}
\arguments{
\item{delta}{The clinical acceptable width of the LOA.}

\item{sigma}{The standard deviation of the differences.}

\item{me}{The mean of the differences. Default is set to 0.}

\item{alpha}{Significance level. Default is set to 0.05.}

\item{beta}{Type II error. Default is set to 0.2.}

\item{gamma}{Coverage of the LOA. Default is set to 0.95.}
}
\value{
Factor variable
}
\description{
Computes the necessary sample size to achieve a given power by way of the formula presented in 'Sample Size for Assessing Agreement between Two Methods of Measurement by Bland-Altman Method' by Lu et al.
}
\examples{
See *my paper*
}
