% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample size for bland altman analysis package.R
\name{Sample size for bland altman analysis package}
\alias{Sample size for bland altman analysis package}
\title{Sample size package for Lin98}
\description{
Contains functions to calculate sample sizes for Bland Altman Analysis
}
\author{
Andreas Kristian Pedersen \email{Andreas.Kristian.Pedersen@rsyd.dk}
}
