% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSizeBAplot.R
\name{SampleSizeBAplot}
\alias{SampleSizeBAplot}
\title{Sample size for Bland Altman analysis}
\description{
Contains functions to calculate sample sizes for Bland Altman Analysis described in the paper 'Sample size consideration for Bland-Altman analysis- a systematic review' by Pedersen et al.
}
\author{
Andreas Kristian Pedersen \email{Andreas.Kristian.Pedersen@rsyd.dk}
}
