% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shieh18assurance.R
\name{Shieh18assurance}
\alias{Shieh18assurance}
\title{Sample size to acquire a given width with a given assurance by way of the confidence interval for the limits of agreement proposed in Shieh's 2018 paper}
\usage{
Shieh18assurance(sigma, ew, ap, alpha = 0.05, me = 0, gamma = 0.95)
}
\arguments{
\item{sigma}{The standard deviation of the differences.}

\item{ew}{The clinical acceptable precision of the confidence interval.}

\item{ap}{Assurance probability.}

\item{alpha}{alpha level of the confidence interval. Default is set to 0.05.}

\item{me}{Mean differences. Default is set to 0.}

\item{gamma}{Coverage of the LOA. Default is set to 0.95.}
}
\value{
Factor variable
}
\description{
Computes the necessary sample size to achieve a given precision with a given assurance probability for the LOA by way of the confidence interval presented in 'The appropriateness of Bland-Altman's approximate confidence intervals for limits of agreement' by Shieh.
}
\examples{
See *my paper*
}
