% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shieh18.R
\name{Shieh18expectedwidth}
\alias{Shieh18expectedwidth}
\title{Sample size to acquire a given width of the confidence interval for the limits of agreement proposed in Shieh's 2018 paper}
\usage{
Shieh18expectedwidth(sigma, ew, gamma = 0.95, alpha = 0.05, mu = 0)
}
\arguments{
\item{sigma}{The standard deviation of the differences.}

\item{ew}{The clinical acceptable width of the confidence interval.}

\item{gamma}{Coverage of the LOA. Default is set to 0.95.}

\item{alpha}{alpha level of the confidence interval. Default is set to 0.05.}

\item{mu}{Mean of the differences.}
}
\value{
Factor variable
}
\description{
Computes the necessary sample size to achieve a given precision for the LOA by way of the confidence interval presented in 'The appropriateness of Bland-Altman's approximate confidence intervals for limits of agreement' by Shieh.
}
\examples{
See *my paper*
}
