% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yi07.R
\name{Yi08}
\alias{Yi08}
\title{Sample size for the agreement test proposed in Yi's 2008 paper}
\usage{
Yi08(delta, sigma, alpha = 0.05, beta = 0.2)
}
\arguments{
\item{delta}{The clinical acceptable variance of the differences.}

\item{sigma}{The standard deviation of the differences.}

\item{alpha}{Significance level. Default is set to 0.05.}

\item{beta}{Type II error. Default is set to 0.2.}
}
\value{
Factor variable
}
\description{
Computes the necessary sample size to achieve a given power by way of the formula presented in 'Reliability analysis for continuous measurements: Equivalence test for agreement' by Yi et al.
}
\examples{
See *my paper*
}
